% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrich_go.R
\name{enrich_go}
\alias{enrich_go}
\title{GO enrichment analysis}
\usage{
enrich_go(nmf, cluster = 5, factor = 20, sp = "org.Hs.eg.db", top = 50)
}
\arguments{
\item{nmf}{The result from function perturbmeta.}

\item{cluster}{The interested cluster for GO enrichment analysis. Default: 5.}

\item{factor}{The number of factor in NMF. Default: 20.}

\item{sp}{The species for gene count matrix. Defalut: org.Hs.eg.db.}

\item{top}{The number of high-contribution features extracted for GO enrichment analysis. Default: 50.}
}
\value{
A list contain the results of Go enrichment analysis for interested clusters.
}
\description{
GO enrichment analysis was performed based on high-contributing genes from the perturbmeta results for interested cluster.
}
\examples{
go_res<-enrich_go(res_nmf, cluster=5, factor=20, sp='org.Hs.eg.db', top=50)
}
