% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_cell.R
\name{filter_cell}
\alias{filter_cell}
\title{Filter cells}
\usage{
filter_cell(
  count,
  sgrna,
  filter = 0.5,
  npcs = 20,
  nfeatures = 3000,
  ctrl = "NT"
)
}
\arguments{
\item{count}{A seurat object that has been normalized.}

\item{sgrna}{A cell-perturbation data frame.}

\item{filter}{Ther ratio of filtering cells that are perturbed non-significant. Default: 0.5.}

\item{npcs}{A number of PCs in PCA. Default: 20 (Same as number of factor)}

\item{nfeatures}{A parameter for extracting high-variable genes. Default: 3000.}

\item{ctrl}{The name of control perturbation. Default: NT}
}
\value{
A filtered seurat object.
}
\description{
Filter the cells that are non-significant perturbed.
}
\examples{
rna_filter<-filter_cell(rna,sgrna,filter=0.5,ctrl='NT')
}
