% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/perturbmeta_score.R
\name{perturbmeta_score}
\alias{perturbmeta_score}
\title{Calculate perturbmeta score for each cluster}
\usage{
perturbmeta_score(nmf, sgrna, cluster = 5, factor = 20, nperms = 1000)
}
\arguments{
\item{nmf}{The result from function perturbmeta.}

\item{sgrna}{Ther cell-perturbation data frame.}

\item{cluster}{The number of clusters that used for perturbmeta.}

\item{factor}{The number of factor in NMF. Default: 20.}

\item{nperms}{The number of permutations. Default: 1000.}
}
\value{
A list contain perturbmeta_score and p value.
}
\description{
Euclidean distance is used to calculate the score of the perturbation to each factor, and then
Monte Carlo permutation is used to test the P-value, and multiple tests are used to correct the P-value.
}
\examples{
score_res<-perturbmeta_score(res_nmf,sgrna,cluster=5,factor=20,nperms=1000)
}
