% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/perturbmeta_score_plot.R
\name{perturbmeta_score_plot}
\alias{perturbmeta_score_plot}
\title{Plot the result of perturbmeta score}
\usage{
perturbmeta_score_plot(score, p, cluster = 5, plot = "all", nfact = 20)
}
\arguments{
\item{score}{The first result of function perturbmeta_score.}

\item{p}{The second result of function perturbmeta_score.}

\item{cluster}{The number of clusters that used for perturbmeta.}

\item{plot}{A parameter for which cluster to plot dotplot. Default: 'all'.}

\item{nfact}{The number of factor in NMF. Default: 20.}
}
\value{
A dotplot for oerturbmeta score from interested cluster.
}
\description{
Use dotplot to plot the result of perturbmeta score. It can plot the significant perturbmeta score (p < 0.05) for all cluster
by setting the plot parameter  to all. While it can plot all perturbmeta score from interested cluster by setting the plot
parameter to the number of interested cluster.
}
\examples{
p<-perturbmeta_score_plot(res_score[[1]], res_score[[2]])
}
