% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RunPLSDA.R
\name{RunPLSDA}
\alias{RunPLSDA}
\title{Run PLS-DA}
\usage{
RunPLSDA(count, sgrna, VIP = 1, ncomp = 20)
}
\arguments{
\item{count}{A count data from selected cluster.}

\item{sgrna}{A cell-perturbation data frame.}

\item{VIP}{The threshold of extracting features from PLS-DA. Default: 1}

\item{ncomp}{A number of component in PLS-DA. Default: 20 (Same as number of factor)}
}
\value{
The list contain cell-embedding, feature-loading and high-contribution genes.
}
\description{
Perform PLS-DA for each cluster, then extract high-contribution genes for further analysis.
}
\examples{
plsda_res<-RunPLSDA(count,sgrna,ncomp=20,VIP=1)
}
